+++ GameObject.hpp
@@ -3,14 +3,14 @@
 
 class GameObject {
 public:
-  GameObject(const char *textureSheet, int x, int y);
+  GameObject(const char *textureSheet, uint8_t x, uint8_t y);
   ~GameObject();
   void Update();
   void Render();
 
 private:
-  int xpos;
-  int ypos;
+  uint8_t xpos;
+  uint8_t ypos;
 
   SDL_Texture *objTexture;
   SDL_Rect srcRect, destRect;
